[db] Move database upgrade logic to separate file
