+++ src/Makefile.am
@@ -71,6 +71,7 @@ forked_daapd_LDADD = -lrt \
 
 forked_daapd_SOURCES = main.c \
 	db.c db.h \
+	db_upgrade.c db_upgrade.h \
 	logger.c logger.h \
 	conffile.c conffile.h \
 	cache.c cache.h \
